Refactor: moved classes to support
